requirements.txt
fastapi[standard]
uvicorn
python-nmap